feat: increase touchedCreate() to removed api /share plain POST.
