+++ src/units/plain/unitTouchedStatus.js
@@ -5,6 +5,9 @@ const path = require('path');
 const winston = require('../../../config/winston.js');
 const {envJSONUnitTouchPath} = require('../../../config/.env.json');
 const _DB_unitsAuthor = require('../../../db/models/index').units_author;
+const {
+  internalError,
+} = require('../../utils/reserrHandler.js');
 
 function _touchedStatus(unitId, userId){
   let rawData = fs.readFileSync(path.join(envJSONUnitTouchPath, "units_touch.json"));
@@ -20,12 +23,12 @@ function _touchedStatus(unitId, userId){
       {touched: length},
       {where: {id_unit: unitId}}
     ).catch((err)=>{
-      throw err;
+      throw new internalError("throw by /units plain GET, touchedStatus check, "+err, 131);
     });
     //then write the new data back to json file
     let modifiedData = JSON.stringify(objTouchedStatus, null, 2);
     fs.writeFile(path.join(envJSONUnitTouchPath, "units_touch.json"), modifiedData, (err) => {
-      if (err) throw err;
+      throw new internalError("throw by /units plain GET, touchedStatus check, "+err, 131);
     });
   }
 }
